city

+++ encoding/models/cfpn.py
@@ -146,8 +146,8 @@ class localUp2(nn.Module):
         src_x = (dest_X+0.5)*scale_w-0.5
         
         # four adjacent point in src
-        src_x_0 = torch.floor(src_x)
-        src_y_0 = torch.floor(src_y)
+        src_x_0 = int(torch.floor(src_x))
+        src_y_0 = int(torch.floor(src_y))
         src_x_1 = torch.where(src_x_0 + 1 < ws - 1, src_x_0 + 1, torch.tensor(ws - 1).view(1,1))
         src_y_1 = torch.where(src_y_0 + 1 < hs - 1, src_y_0 + 1, torch.tensor(hs - 1).view(1,1))
